refactored jsx, css files in their respective folders according to page
