+++ src/HomePage/HomePage.jsx
@@ -1,0 +1,17 @@
+import HomeHero from './HomeHero'
+import StatsSection from '../components/StatsSection'
+import MissionSection from '../components/MissionSection'
+import WhoWeAreSection from '../components/WhoWeAreSection'
+import LastScreenSection from '../components/LastScreenSection'
+
+const HomePage = () => (
+  <>
+    <HomeHero />
+    <StatsSection />
+    <MissionSection />
+    <WhoWeAreSection />
+    <LastScreenSection />
+  </>
+)
+
+export default HomePage
