--- src/InfoPage.css
@@ -1,321 +1,0 @@
-.info-page {
-  --info-page-image-ratio: 0.5;
-  --info-page-fluid-step: clamp(
-    0rem,
-    calc((clamp(320px, 100vw, 1440px) - 320px) * 0.0035),
-    4rem
-  );
-
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: max(
-    calc(100vh - var(--nav-height, 0px)),
-    calc(100vw * var(--info-page-image-ratio))
-  );
-  padding: clamp(4rem, 8vw, 6.5rem) clamp(1.5rem, 6vw, 5rem);
-  text-align: center;
-  overflow: hidden;
-  color: #ffffff;
-}
-
-.info-page::before {
-  content: '';
-  position: absolute;
-  inset: 0;
-  background: none;
-  z-index: 0;
-}
-
-.info-page--recruitment {
-  --info-page-image-ratio: 0.436;
-  background-color: #050814;
-  background-image: url('./assets/pages/projects/hero-background.png');
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: contain;
-}
-
-.info-page--sponsorship {
-  --info-page-image-ratio: 0.746;
-  background: url('./assets/pages/sponsorship/hero-background.png') center / cover no-repeat;
-}
-
-.info-page__content {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: clamp(1.25rem, 3vw, 2rem);
-  max-width: 40rem;
-  z-index: 2;
-}
-
-
-.info-page__eyebrow {
-  margin: 0;
-  font-size: clamp(0.9rem, 2vw, 1.2rem);
-  letter-spacing: 0.4em;
-  text-transform: uppercase;
-  color: rgba(255, 255, 255, 0.72);
-}
-
-.info-page__title {
-  margin: 0;
-  font-size: clamp(3rem, 10vw, 5rem);
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-  font-weight: 800;
-}
-
-.info-page__body {
-  margin: 0;
-  font-size: clamp(1.05rem, 2.6vw, 1.4rem);
-  line-height: 1.6;
-  color: rgba(255, 255, 255, 0.85);
-}
-
-.info-page__cta {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  align-self: center;
-  min-width: 12rem;
-  padding: 0.85rem 2.5rem;
-  border-radius: 999px;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  color: #050814;
-  background-color: #4b6fff;
-  text-decoration: none;
-  transition: transform 0.2s ease, filter 0.2s ease;
-}
-
-.info-page__cta:hover,
-.info-page__cta:focus-visible {
-  transform: translateY(-1px);
-  filter: brightness(1.05);
-  color: #ffffff;
-}
-
-.sponsorship-details {
-  background-color: #040a14;
-  padding: clamp(3.5rem, 8vw, 6rem) clamp(1.5rem, 8vw, 5rem) clamp(4rem, 10vw, 7rem);
-  color: rgba(236, 242, 255, 0.95);
-  position: relative;
-  overflow: hidden;
-}
-
-.sponsorship-details__container {
-  max-width: 60rem;
-  margin: 0 auto;
-  display: grid;
-  gap: clamp(1.5rem, 3vw, 2.5rem);
-  text-align: left;
-  position: relative;
-  z-index: 1;
-}
-
-.sponsorship-details__title {
-  margin: 0;
-  font-size: clamp(2.4rem, 1.6rem + 2.6vw, 3.6rem);
-  text-transform: uppercase;
-  letter-spacing: clamp(0.18em, 0.1em + 0.24vw, 0.32em);
-  font-weight: 800;
-  color: #ffffff;
-}
-
-.sponsorship-details__body {
-  margin: 0;
-  font-size: clamp(1rem, 0.95rem + 0.55vw, 1.35rem);
-  line-height: 1.75;
-  letter-spacing: 0.03em;
-  color: rgba(224, 236, 255, 0.88);
-}
-
-.sponsorship-details__actions {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: clamp(1rem, 3vw, 1.75rem);
-  margin-top: clamp(1.5rem, 4vw, 2.5rem);
-}
-
-.sponsorship-details__cta {
-  align-self: center;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0.85rem 2.75rem;
-  border-radius: 999px;
-  background: linear-gradient(135deg, #2da0dc 0%, #6ae0ff 100%);
-  color: #041121;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  font-size: 0.95rem;
-  text-decoration: none;
-  box-shadow: 0 18px 38px rgba(45, 160, 220, 0.35);
-  transition: transform 0.2s ease, box-shadow 0.2s ease, filter 0.2s ease;
-}
-
-.sponsorship-details__cta:hover,
-.sponsorship-details__cta:focus-visible {
-  transform: translateY(-2px);
-  box-shadow: 0 22px 44px rgba(45, 160, 220, 0.45);
-  filter: brightness(1.05);
-  color: #ffffff;
-}
-
-.sponsorship-details__cta--secondary {
-  background: transparent;
-  border: 2px solid rgba(224, 236, 255, 0.85);
-  color: rgba(224, 236, 255, 0.92);
-  box-shadow: none;
-}
-
-.sponsorship-details__cta--secondary:hover,
-.sponsorship-details__cta--secondary:focus-visible {
-  background: rgba(224, 236, 255, 0.12);
-  filter: none;
-  box-shadow: 0 12px 28px rgba(45, 160, 220, 0.25);
-  color: #ffffff;
-}
-
-.sponsor-tiers {
-  background: radial-gradient(circle at top, rgba(35, 72, 132, 0.25), transparent 60%),
-    #03070f;
-  padding: clamp(3rem, 7vw, 6rem) clamp(1.5rem, 8vw, 5rem) clamp(3.5rem, 9vw, 6rem);
-  color: rgba(236, 242, 255, 0.95);
-}
-
-.sponsor-tiers__container {
-  max-width: 72rem;
-  margin: 0 auto;
-  display: grid;
-  gap: clamp(2rem, 4vw, 3rem);
-  text-align: center;
-}
-
-.sponsor-tiers__title {
-  margin: 0;
-  font-size: clamp(2rem, 1.2rem + 3vw, 3.4rem);
-  letter-spacing: 0.12em;
-  text-transform: uppercase;
-  color: #ffffff;
-}
-
-.sponsor-tiers__tier {
-  display: grid;
-  gap: clamp(1rem, 3vw, 1.75rem);
-}
-
-.sponsorship-details__drone {
-  display: none;
-}
-
-@media (min-width: 1024px) {
-  .sponsorship-details__drone {
-    display: block;
-    position: absolute;
-    top: clamp(5.5rem, 4vw, 6.5rem);
-    left: 50%;
-    width: clamp(420px, 38vw, 640px);
-    height: clamp(240px, 22vw, 380px);
-    transform: translateX(-50%) rotate(-8deg);
-    background: url('./assets/home/last-screen/drone.png') center/contain no-repeat;
-    opacity: 0.38;
-    filter: drop-shadow(0 24px 48px rgba(0, 0, 0, 0.55));
-    pointer-events: none;
-    z-index: 2;
-  }
-
-  .sponsorship-details__body,
-  .sponsorship-details__actions,
-  .sponsorship-details__cta {
-    position: relative;
-    z-index: 1;
-  }
-}
-
-.sponsor-tiers__tier-title {
-  margin: 0;
-  font-size: clamp(1.35rem, 1rem + 1.5vw, 2rem);
-  letter-spacing: 0.18em;
-  text-transform: uppercase;
-  color: rgba(238, 245, 255, 0.9);
-}
-
-.sponsor-tiers__logos {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-  gap: clamp(1.5rem, 4vw, 3rem);
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.sponsor-tiers__link {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  padding: clamp(0.75rem, 2.2vw, 1.1rem) clamp(1.25rem, 3vw, 2rem);
-  border-radius: 24px;
-  background-color: rgba(255, 255, 255, 0.92);
-  border: 1px solid rgba(255, 255, 255, 0.45);
-  transition: transform 0.2s ease, filter 0.2s ease, box-shadow 0.2s ease;
-}
-
-.sponsor-tiers__link:hover,
-.sponsor-tiers__link:focus-visible {
-  transform: translateY(-3px);
-  filter: none;
-  box-shadow: 0 22px 44px rgba(24, 110, 184, 0.25);
-  color: #ffffff;
-}
-
-.sponsor-tiers__image {
-  display: block;
-  width: clamp(6.5rem, 3.5rem + 8vw, 13rem);
-  height: auto;
-  max-height: clamp(3.75rem, 2rem + 3.6vw, 5.5rem);
-}
-
-.sponsor-tiers__logos--title .sponsor-tiers__image {
-  width: clamp(8rem, 4.5rem + 10vw, 15rem);
-  max-height: clamp(4.25rem, 2.8rem + 3.6vw, 6.5rem);
-}
-
-.sponsor-tiers__logos--gold .sponsor-tiers__image {
-  width: clamp(6rem, 3.25rem + 7vw, 11rem);
-  max-height: clamp(3.4rem, 1.8rem + 2.8vw, 4.75rem);
-}
-
-.sponsor-tiers__logos--silver .sponsor-tiers__image {
-  width: clamp(5.5rem, 3rem + 6vw, 10rem);
-  max-height: clamp(3.1rem, 1.6rem + 2.4vw, 4.25rem);
-}
-
-.sponsor-tiers__logos--bronze .sponsor-tiers__image {
-  width: clamp(5rem, 2.8rem + 5.5vw, 9rem);
-  max-height: clamp(3rem, 1.6rem + 2vw, 4rem);
-}
-
-.sponsor-tiers__logos--contributor .sponsor-tiers__image {
-  width: clamp(4.75rem, 2.6rem + 5vw, 8.5rem);
-  max-height: clamp(2.9rem, 1.5rem + 1.9vw, 3.8rem);
-}
-
-@media (max-width: 600px) {
-  .info-page {
-    padding-top: 6rem;
-    padding-bottom: 6rem;
-  }
-
-  .info-page__body {
-    font-size: clamp(1rem, 4vw, 1.25rem);
-  }
-}
