--- src/RecruitmentPage.css
@@ -1,330 +1,0 @@
-/* RecruitmentPage.css
-   Styles for the "Fall Recruitment Cycle" section that sits
-   BELOW the main Recruitment hero.
-*/
-
-/* Wrapper section under the hero */
-.recruitment-details {
-    background: #050814; /* match site background */
-    padding: 4.5rem 0 5.5rem;
-    display: flex;
-    justify-content: center;
-  }
-  
-  /* Main card */
-  .recruitment-panel {
-    width: 100%;
-    max-width: 1080px;
-    padding: 2.5rem 3.25rem 3rem;
-    border-radius: 1.8rem;
-    background: rgba(6, 10, 22, 0.97);
-    backdrop-filter: blur(18px);
-    box-shadow: 0 20px 48px rgba(0, 0, 0, 0.6);
-    color: #f5f7ff;
-    text-align: center;
-  }
-  
-  /* Header / CTA */
-  
-  .recruitment-panel__header {
-    margin-bottom: 2.3rem;
-  }
-  
-  .recruitment-panel__title-main {
-    font-size: 2.2rem;
-    font-weight: 700;
-    letter-spacing: 0.14em;
-    text-transform: uppercase;
-    margin-bottom: 0.9rem;
-  }
-  
-  .recruitment-panel__body {
-    font-size: 1.5rem;
-    line-height: 1.6;
-    max-width: 45rem;
-    color: #ccd3e5;
-    margin: 0 auto 1.6rem;
-  }
-  
-  .recruitment-panel__cta {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0.6rem 1.8rem;
-    border-radius: 999px;
-    font-size: 0.86rem;
-    font-weight: 600;
-    background: #2f89ff;
-    color: #ffffff;
-    text-decoration: none;
-    transition:
-      background 0.15s ease,
-      transform 0.15s ease,
-      box-shadow 0.15s ease;
-  }
-  
-  .recruitment-panel__cta:hover {
-    background: #3d98ff;
-    transform: translateY(-1px);
-    box-shadow: 0 10px 26px rgba(0, 0, 0, 0.4);
-  }
-  
-  /* Section headings inside the card (Timeline / Team Info) */
-  
-  .recruitment-section {
-    margin-top: 2.2rem;
-  }
-  
-  .recruitment-section__heading {
-    font-size: 0.95rem;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 0.16em;
-    margin-bottom: 1.5rem;
-    color: #aeb7cc;
-    align-self: center;
-  }
-  
-  /* make "Recruitment Timeline" closer to the Figma */
-  .recruitment-section__heading--timeline {
-    font-size: 1.1rem;
-    letter-spacing: 0.18em;
-    text-align: center;
-    margin-bottom: 4rem;
-  }
-  
-  /* === CURVED BLUE TIMELINE === */
-  
-  .recruitment-timeline {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  /* Match Figma-ish aspect ratio for the curve */
-  .recruitment-timeline__canvas {
-    position: relative;
-    width: min(100%, 800px);
-    margin: 0 auto;
-    aspect-ratio: 600 / 360;
-  }
-  
-  /* SVG curve */
-  /* SVG curve */
-.recruitment-timeline__path {
-    position: absolute;
-    inset: 0;
-    width: 100%;
-    height: 100%;
-    pointer-events: none;
-  }
-  
-  /* Base dashed style */
-  .recruitment-timeline__path-line {
-    fill: none;
-    stroke: #2da0dc;          /* Figma blue */
-    stroke-width: 2px;
-    stroke-dasharray: 10 30;  /* Figma dashes: 10, 30 */
-    stroke-linecap: round;
-  }
-  
-  /* Animation layer – applies because of the extra class in JSX */
-  .recruitment-timeline__path-line--animated {
-    /* Big dash/offset to "hide" the line initially */
-    stroke-dasharray: 1000;
-    stroke-dashoffset: 1000;
-    animation:
-      /* 0s → 2s: your original wipe-in */
-      timeline-draw 2s ease-out forwards,
-  
-      /* at 2s: snap to dashed pattern */
-      timeline-dashify 0.01s linear 2s forwards,
-  
-      /* 2s → 4.2s: gentle start of motion */
-      timeline-soft-scroll 2.2s ease-out 2s forwards,
-  
-      /* from 4.2s onward: steady conveyor belt, looping */
-      timeline-scroll 4s linear 4.2s infinite;
-  }
-  
-  /* 1) Draw the line from start to end (same as you had) */
-  @keyframes timeline-draw {
-    from {
-      stroke-dashoffset: 1000;
-    }
-    to {
-      stroke-dashoffset: 0;
-    }
-  }
-  
-  /* 2) After it’s fully drawn, switch to dashed pattern (same as you had) */
-  @keyframes timeline-dashify {
-    to {
-      stroke-dasharray: 10 30;
-    }
-  }
-  
-  /* 3) Gentle “fade into motion” once, right after the draw */
-  @keyframes timeline-soft-scroll {
-    from {
-      stroke-dashoffset: 0;
-    }
-    to {
-      stroke-dashoffset: -20;  /* small move, looks like easing into motion */
-    }
-  }
-  
-  /* 4) Then normal conveyor belt forever */
-  @keyframes timeline-scroll {
-    from {
-      stroke-dashoffset: -20;
-    }
-    to {
-      stroke-dashoffset: -60;  /* same direction, steady speed */
-    }
-  }
-  
-  
-  .recruitment-timeline__icon,
-  .recruitment-timeline__flag-icon {
-    pointer-events: none;
-  }
-  
-  /* Text / events overlay */
-  .recruitment-timeline__list {
-    position: absolute;
-    inset: 0;
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    font-size: 0.88rem;
-    color: #f5f7ff;
-  }
-  
-  .recruitment-timeline__item {
-    position: absolute;
-    max-width: 260px;
-  }
-  
-  /* Positions tuned to match Figma and stay off the line */
-  .recruitment-timeline__item--imagine {
-    top: 15%;
-    right: 58%;
-    text-align: right;
-  }
-  
-  .recruitment-timeline__item--apps-open {
-    top: -10%;
-    left: 7%;
-    text-align: left;
-  }
-  
-  .recruitment-timeline__item--in-person {
-    top: 9%;
-    left: 55%;
-    text-align: left;
-  }
-  
-  .recruitment-timeline__item--online {
-    top: 38%;
-    right: -7%;
-    text-align: center; /* 'middle' isn't valid; use center */
-  }
-  
-  .recruitment-timeline__item--close {
-    top: 62%;
-    right: 24%;
-    text-align: left;
-  }
-  
-  .recruitment-timeline__item--results {
-    top: 68%;
-    right: 60%;
-    text-align: left;
-  }
-  
-  .recruitment-timeline__item--meeting {
-    top: 92%;
-    left: 8%;
-    text-align: left;
-  }
-  
-  /* Typography */
-  .recruitment-timeline__label {
-    font-family: "Hiragino Kaku Gothic ProN", "Hiragino Kaku Gothic Pro",
-      system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-    font-weight: 600;
-    font-size: 1rem;
-    line-height: 0.6;
-    letter-spacing: 0.03em;
-    color: #ffffff;
-  }
-  
-  .recruitment-timeline__meta {
-    font-size: 0.78rem;
-    color: #b7c0d3;
-  }
-  
-  /* === TEAM INFORMATION === */
-  
-  .recruitment-teams__intro {
-    font-size: 1.2rem;
-    line-height: 1.5;
-    color: #c7cfe0;
-    max-width: 45rem;
-    margin: 0.25rem auto 2rem;
-    text-align: center;
-  }
-  
-  .recruitment-teams__grid {
-    display: grid;
-    grid-template-columns: repeat(2, minmax(0, 1fr)); /* 2×2 grid on desktop */
-    gap: 3rem 4rem;
-    justify-items: center;
-  }
-  
-  .recruitment-team {
-    max-width: 300px;
-    text-align: center;
-  }
-  
-  .recruitment-team__name {
-    font-size: 1.25rem;
-    font-weight: 700;
-    margin-bottom: 0.4rem;
-  }
-  
-  .recruitment-team__body {
-    font-size: 1.1rem;
-    line-height: 1.5;
-    color: #c5cde0;
-  }
-  
-  /* === MOBILE: fall back to vertical list === */
-  
-  @media (max-width: 768px) {
-    .recruitment-timeline__canvas {
-      aspect-ratio: auto;
-    }
-  
-    .recruitment-timeline__path {
-      display: none;
-    }
-  
-    .recruitment-timeline__list {
-      position: static;
-      padding-top: 0.5rem;
-    }
-  
-    .recruitment-timeline__item {
-      position: static;
-      max-width: none;
-      margin-bottom: 1.5rem;
-      text-align: left;
-    }
-  
-    .recruitment-teams__grid {
-      grid-template-columns: 1fr;
-    }
-  }
-
