+++ src/SponsorshipPage/SponsorshipPage.jsx
@@ -1,18 +1,18 @@
-import sponsorshipPackage from './assets/documents/UAS-Sponsorship-Package-2025-2026.pdf'
-import apscLogo from './assets/pages/sponsorship/apsc-logo.webp'
-import onshapeLogo from './assets/pages/sponsorship/onshape-colour.svg'
-import altiumLogo from './assets/pages/sponsorship/altium-black.svg'
-import atlassianLogo from './assets/pages/sponsorship/atlassian-blue.svg'
-import coastalDroneLogo from './assets/pages/sponsorship/coastal-drone.webp'
-import walterGageLogo from './assets/pages/sponsorship/walter-gage.svg'
-import cubepilotLogo from './assets/pages/sponsorship/cubepilot.webp'
-import zaberLogo from './assets/pages/sponsorship/zaber.webp'
-import dockerLogo from './assets/pages/sponsorship/docker-blue.svg'
-import simnetLogo from './assets/pages/sponsorship/simnet.webp'
-import stickerMuleLogo from './assets/pages/sponsorship/sticker-mule.svg'
-import rockWestCompositesLogo from './assets/pages/sponsorship/rock-west-composites.webp'
-import actuonixLogo from './assets/pages/sponsorship/actuonix.webp'
-import './InfoPage.css'
+import sponsorshipPackage from '../assets/documents/UAS-Sponsorship-Package-2025-2026.pdf'
+import apscLogo from '../assets/pages/sponsorship/apsc-logo.webp'
+import onshapeLogo from '../assets/pages/sponsorship/onshape-colour.svg'
+import altiumLogo from '../assets/pages/sponsorship/altium-black.svg'
+import atlassianLogo from '../assets/pages/sponsorship/atlassian-blue.svg'
+import coastalDroneLogo from '../assets/pages/sponsorship/coastal-drone.webp'
+import walterGageLogo from '../assets/pages/sponsorship/walter-gage.svg'
+import cubepilotLogo from '../assets/pages/sponsorship/cubepilot.webp'
+import zaberLogo from '../assets/pages/sponsorship/zaber.webp'
+import dockerLogo from '../assets/pages/sponsorship/docker-blue.svg'
+import simnetLogo from '../assets/pages/sponsorship/simnet.webp'
+import stickerMuleLogo from '../assets/pages/sponsorship/sticker-mule.svg'
+import rockWestCompositesLogo from '../assets/pages/sponsorship/rock-west-composites.webp'
+import actuonixLogo from '../assets/pages/sponsorship/actuonix.webp'
+import './SponsorshipPage.css'
 
 const SponsorshipPage = () => {
   return (
